requirements.txt
flask
openai-whisper
torch
librosa
numpy
soundfile
webrtcvad
pydub
scipy
noisereduce
redis
